Crate pallet_xcm_weight_trader

A pallet to trade weight for XCM execution

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Structs

Trader

XcmWeightTraderAssetInfo

Constants

RELATIVE_PRICE_DECIMALS

Functions

compute_fee_amount

Helper function to compute fee amount from weight and asset location. This is used by the XcmTransactorFeeTrader adapter implementation.