pallet_xcm_weight_trader::pallet
Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

SupportedAssets

Stores all supported assets per XCM Location. The u128 is the asset price relative to native asset with 18 decimals The boolean specify if the support for this asset is active