Trait pallet_xcm_weight_trader::pallet::Config

pub trait Config: Config {

    type AccountIdToLocation: Convert<Self::AccountId, Location>;
    type AddSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type AssetLocationFilter: Contains<Location>;
    type AssetTransactor: TransactAsset;
    type Balance: TryInto<u128>;
    type EditSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type NativeLocation: Get<Location>;
    type PauseSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RemoveSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type ResumeSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type WeightInfo: WeightInfo;
    type WeightToFee: WeightToFee<Balance = Self::Balance>;
    type XcmFeesAccount: Get<Self::AccountId>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configuration trait of this pallet.

Required Associated Types

type AccountIdToLocation: Convert<Self::AccountId, Location>

Convert T::AccountId to Location.

type AddSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to register a supported asset

type AssetLocationFilter: Contains<Location>

A filter to forbid some XCM Location to be supported for fees. if you don’t use it, put “Everything”.

type AssetTransactor: TransactAsset

How to withdraw and deposit an asset.

type Balance: TryInto<u128>

The native balance type.

type EditSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to edit a supported asset units per seconds

type NativeLocation: Get<Location>

XCM Location for native curreny

type PauseSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to pause a supported asset

type RemoveSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to remove a supported asset

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type ResumeSupportedAssetOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to unpause a supported asset

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type WeightToFee: WeightToFee<Balance = Self::Balance>

Convert a weight value into deductible native balance.

type XcmFeesAccount: Get<Self::AccountId>

Account that will receive xcm fees

Object Safety

This trait is not object safe.